requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
category==0.1.0
contextlib2==21.6.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphql-relay==2.0.1
numpy==1.22.3
promise==2.3
PyJWT==1.7.1
pymemcache==3.5.1
python-memcached==1.59
rust-category==0.2.0
Rx==1.6.1
schema==0.7.5
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3


graphene>=2.1,<3
graphene-django>=2.1,<3
graphql-core>=2.1,<3
django>=3.1.14,<4
django-filter>=2
python-dateutil
